Update drawer and split out 3 components
