+++ src/App.js
@@ -14,41 +14,28 @@ import './mediaQuery.css';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import AppBar from '@material-ui/core/AppBar';
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
+
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import { withStyles } from '@material-ui/core/styles';
 import { Link } from 'react-scroll';
-
+import FlippingIcons from './FlippingIcons.js';
 
 // Images for the page --------------------------------------------------------
-import quiz_screenshot from './images/quiz_screenshot.png';
-import amp_library_screenshot from './images/amp-library-screenshot.jpg';
-import bug_tracker_screenshot from './images/bug_tracker_screenshot.png';
-import drum_machine_screenshot from './images/drum-machine-screenshot.jpg';
-import maze_game_screenshot from './images/maze_game_screenshot.png';
-import weather_screenshot from './images/weather_app.png';
 import logo from './images/Adam_logo.png';
 import skyline from './images/skyline.jpg';
 import pin from './images/maps-pin-smaller.png';
-import /* webpackPrefetch: true */ html from './images/html.png';
-import /* webpackPrefetch: true */ css from './images/css.png';
-import /* webpackPrefetch: true */ js from './images/js.png';
-import /* webpackPrefetch: true */ git from './images/git.png';
-import /* webpackPrefetch: true */ react from './images/react-1.png';
-import /* webpackPrefetch: true */ node from './images/node-js-icon.png';
 import /* webpackPrefetch: true */ github from './images/github-icon.png';
-import /* webpackPrefetch: true */ mysql from './images/mysql-icon.png';
 import /* webpackPrefetch: true */ linkedin from './images/linkedin-icon.png';
-import /* webpackPrefetch: true */ java from './images/java_icon.png';
 
-// Lazy Imports
-const TemporaryDrawer = lazy(() => import('./temporaryDrawer.js'));
+// Lazy Imports ---------------------------------------------------------------
+const TemporaryDrawer = lazy(() => import('./TemporaryDrawer.js'));
+const SmallerProjects = lazy(() => import('./SmallerProjects.js'));
+const LargerProjects = lazy(() => import('./LargerProjects.js'));
 const renderLoader = () => <p>Loading</p>;
+
 
 // Material UI styles object --------------------------------------------------
 const styles = theme => ({
@@ -108,154 +95,7 @@ const styles = theme => ({
 });
 
 
-// Array of objects holding the data for grid of tech skills icons below ------
-const TECH_ICONS_DATA = [
-    {
-        href: 'https://en.wikipedia.org/wiki/HTML',
-        className: 'tech-icon',
-        id: 'html-icon',
-        alt: 'html',
-        src: html,
-        width: "166",
-        height: "166"
-    },
-    {
-        href: 'https://en.wikipedia.org/wiki/Cascading_Style_Sheets',
-        className: 'tech-icon',
-        id: 'css-icon',
-        alt: 'css',
-        src: css,
-        width: "160",
-        height: "160"
-    },
-    {
-        href: 'https://en.wikipedia.org/wiki/JavaScript',
-        className: 'tech-icon',
-        id: 'js-icon',
-        alt: 'js',
-        src: js,
-        width: "155",
-        height: "155"
-    },
-    {
-        href: 'https://en.wikipedia.org/wiki/React_(JavaScript_library)',
-        className: 'tech-icon',
-        id: 'react-icon',
-        alt: 'react',
-        src: react,
-        width: "161",
-        height: "161"
-    },
-    {
-        href: 'https://en.wikipedia.org/wiki/Java_(programming_language)',
-        className: 'tech-icon',
-        id: 'java-icon',
-        alt: 'Java',
-        src: java,
-        width: "162",
-        height: "163"
-    },
-    {
-        href: 'https://en.wikipedia.org/wiki/Node.js',
-        className: 'tech-icon',
-        id: 'node-icon',
-        alt: 'node',
-        src: node,
-        width: "192",
-        height: "192"
-    },
-    {
-        href: 'https://en.wikipedia.org/wiki/MySQL',
-        className: 'tech-icon',
-        id: 'mysql-icon',
-        alt: 'mysql',
-        src: mysql,
-        width: "162",
-        height: "162"
-    },
-    {
-        href: 'https://en.wikipedia.org/wiki/Git',
-        className: 'tech-icon',
-        id: 'git-icon',
-        alt: 'git',
-        src: git,
-        width: "131",
-        height: "131"
-    }
-];
-
-
-// Object holding the information for the two upper project's cards -----------
-const LARGE_PROJ_CARDS = [
-    {
-        key: 1,
-        image: quiz_screenshot,
-        title: 'QuizSoft Quiz Builder',
-        project: 'QuizSoft Quiz Builder',
-        link: 'https://adamjwright.com/quiz_soft',
-        about:
-            `Node.js / MongoDB / Handlebars - Quiz builder where the user can create a set of
-            quizzes and then email a one time link and see the results graded and ranked automatically.`
-    },
-    {
-        key: 2,
-        image: bug_tracker_screenshot,
-        title: 'Software Bug Tracker',
-        project: 'Software Bug Tracker',
-        link: 'https://adamjwright.com/bug_tracker/login',
-        about:
-            `Java / Spring Boot / MySQL / Maven - An application which allows the user to add new 
-            software bugs and to track and update their status as they are fixed.`
-        
-    }
-];
-
-
-// Object holding the information for the four lower project's cards ----------
-const SMALL_PROJ_CARDS = [
-    {
-        key: 1,
-        image: amp_library_screenshot,
-        title: 'Amp Information Library',
-        project: 'Amp Information Library',
-        link: 'https://adamjwright.com/amp_library',
-        about:
-            `React.js / Firebase Realtime DB - App where a user can add photos, schematics, 
-            settings, and details about electric guitar amplifier circuits.`
-    },
-    {
-        key: 2,
-        image: maze_game_screenshot,
-        title: 'C++ WASM Maze Game',
-        project: 'C++ WASM Maze Game',
-        link: 'https://adamjwright.com/wasm_maze_game',
-        about:
-            `C++ / Web Assembly / JavaScript - Command line maze game converted to run in the browser using
-            emscripten and an HTML and CSS CLI.`
-    },
-    {
-        key: 3,
-        image: drum_machine_screenshot,
-        title: 'React Drum Machine',
-        project: 'React Drum Machine',
-        link: 'https://adamjwright.com/drum_machine',
-        about:
-            `React.js - Drum machine that can be played by either clicking its buttons or pressing 
-            the keyboard keys shown.`
-    },
-    {
-        key: 4,
-        image: weather_screenshot,
-        title: 'Weather Widget',
-        project: 'Weather Widget',
-        link: 'https://adamjwright.com/weather_widget',
-        about:
-            `JavaScript - Weather widget which finds the user's location and shows the local weather and
-            allows the user to manually change locations.`
-    }
-];
-
-// Page text ------------------------------------------------------------------
+// About Me text --------------------------------------------------------------
 const text = {
     about: `My name is Adam Wright and I'm a software engineer for Open Sky Software.
     I live in Seattle and I'm an Oregon State University alumni with a B.S. in 
@@ -266,7 +106,6 @@ const text = {
     email: `You can reach out to me here on the contact page or email me at 
     adamjw321@gmail.com.`
 }
-
 
 // Page layout ----------------------------------------------------------------
 function Album(props) {
@@ -340,19 +179,7 @@ function Album(props) {
                             <Grid id="icon-container" item xs={9} sm={9} md={9} lg={12}>
 
                                 {/* Grid of tech skills icons */}
-
-                                {TECH_ICONS_DATA.map(data => (
-                                    <a key={data.id} href={data.href}>
-                                        <img
-                                            className={data.className}
-                                            id={data.id}
-                                            alt={data.alt}
-                                            src={data.src}
-                                            width={data.width}
-                                            height={data.height}
-                                        />
-                                    </a>
-                                ))}
+                                <FlippingIcons />
 
                             </Grid>
                         </Grid>
@@ -365,49 +192,20 @@ function Album(props) {
                     <div className="full-stack" id="stack">
                         <Grid container spacing={5} justify="center">
 
-                            {/* This section maps the LARGE_PROJ_CARDS object into 2 card components */}
-                            {LARGE_PROJ_CARDS.map(card => (
-                                <Grid item key={card.key} sm={6} md={5} lg={5}>
-                                    <Card className={classes.card}>
-                                        <a href={card.link}>
-                                            <CardMedia 
-                                                className={classes.cardMedia} 
-                                                image={card.image} 
-                                                title={card.title}
-                                            />
-                                        </a>
-                                        <CardContent className={classes.cardContent}>
-                                            <Typography gutterBottom variant="h5" component="h2"> {card.project} </Typography>
-                                            <Typography> {card.about} </Typography>
-                                        </CardContent>
-                                    </Card>
-                                </Grid>
-                            ))}
+                            <Suspense fallback={renderLoader()}>
+                                <LargerProjects classes={classes} />
+                            </Suspense>
+
                         </Grid>
                     </div>
 
-                    {/* Smaller project cards ----------------------------- */}
+                    {/* 4 Smaller project cards --------------------------- */}
                     <div className="smaller-apps">
                         <Grid container spacing={5}>
-
-                            {/* This section maps the SMALL_PROJ_CARDS object into 4 card components */}
-                            {SMALL_PROJ_CARDS.map(card => (
-                                <Grid item key={card.key} sm={6} md={3} lg={3}>
-                                    <Card className={classes.card}>
-                                        <a href={card.link}>
-                                            <CardMedia 
-                                                className={classes.cardMedia} 
-                                                image={card.image} 
-                                                title={card.title}
-                                            />
-                                        </a>
-                                        <CardContent className={classes.cardContent}>
-                                            <Typography gutterBottom variant="h5" component="h2"> {card.project} </Typography>
-                                            <Typography> {card.about} </Typography>
-                                        </CardContent>
-                                    </Card>
-                                </Grid>
-                            ))}
+                            
+                            <Suspense fallback={renderLoader()}>
+                                <SmallerProjects classes={classes} />
+                            </Suspense>
 
                         </Grid>
                     </div>
@@ -418,7 +216,6 @@ function Album(props) {
             {/* Footer, which is the bottom 1/3 of the page --------------- */}
             <footer id="footer" className={classes.footer}>
                 <div className="bottom-container">
-
 
                     {/* About me box -------------------------------------- */}
                     <div id="about-text">
@@ -431,7 +228,6 @@ function Album(props) {
                             {text.email}
                         </Typography> 
                     </div>
-
 
                     {/* Social Links -------------------------------------- */}
                     <div id="social">
@@ -462,7 +258,6 @@ function Album(props) {
                             </Typography>
                         </a>
                     </div>
-
 
                     {/* The skyline background and location pin ----------- */}
                     <img
